Move all CTF parsing to hansei-types and use it in hansei
