+++ hansei-types/Cargo.toml
@@ -7,7 +7,10 @@ edition = "2024"
 anyhow.workspace = true
 debugdb.workspace = true
 derive_more.workspace = true
+durin.workspace = true
 humantime.workspace = true
 proc.workspace = true
+regex.workspace = true
 reify.workspace = true
 rustc-demangle.workspace = true
+semver.workspace = true
